Moved the function to LocatorController and fixed nullpointer, graphnodes are clickable again.
