+++ geex-core/src/main/java/nl/tudelft/context/controller/locator/LocatorController.java
@@ -2,18 +2,17 @@ package nl.tudelft.context.controller.locator;
 
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.scene.Node;
 import javafx.scene.layout.Pane;
 import javafx.scene.shape.Rectangle;
 import nl.tudelft.context.controller.AbstractGraphController;
+import nl.tudelft.context.drawable.DrawableLocatorMutation;
 import nl.tudelft.context.drawable.graph.AbstractDrawableNode;
 import nl.tudelft.context.model.graph.DefaultNode;
+import nl.tudelft.context.model.graph.GraphNode;
+import nl.tudelft.context.model.graph.StackGraph;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 /**
@@ -156,7 +155,34 @@ public class LocatorController {
 
         if (width != locator.getWidth()) {
             width = locator.getWidth();
-            graphController.showMutationsInLocator();
+            showMutationsInLocator();
+        }
+
+    }
+
+    /**
+     * The function that draws the mutations in the positionbar.
+     */
+    public void showMutationsInLocator() {
+
+        StackGraph currentGraph = graphController.getCurrentGraph();
+
+        if(currentGraph != null) {
+            Set<DefaultNode> mutations = currentGraph.vertexSet().stream()
+                    .filter(node -> node instanceof GraphNode).collect(Collectors.toSet());
+
+            int max = currentGraph.vertexSet().stream()
+                    .mapToInt(DefaultNode::getRefEndPosition)
+                    .max().getAsInt();
+
+
+            Node locatorIndicator = locator.getChildren().get(0);
+            locator.getChildren().clear();
+            locator.getChildren().add(locatorIndicator);
+
+            mutations.forEach(node -> locator
+                    .getChildren()
+                    .add(new DrawableLocatorMutation(node, locator.getWidth(), max)));
         }
 
     }
